feat: implement GitHub webhook signature verification and raw body parsing
